requirements.txt
# Core HTTP and HTML parsing
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Data validation and modeling
pydantic>=2.5.0

# Date parsing
python-dateutil>=2.8.2

# Retry and rate limiting
tenacity>=8.2.0

# Type hints
typing-extensions>=4.8.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
vcrpy>=4.4.0
